requirements.txt
anthropic
pandas
requests
python-dotenv
openpyxl
reportlab
streamlit
